requirements.txt
black==19.10b
PyYAML==5.2
requests==2.22.0
xmltodict==0.12.0